# backend/docs/swagger.yaml
definitions:
  errors.ErrorInfo:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: 请求参数错误
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorInfo'
      success:
        example: false
        type: boolean
    type: object
...
  models.User:
    properties:
      avatar:
        description: 头像 URL
        type: string
      balance:
        description: 余额
        type: number
      created_at:
        description: 创建时间
        type: string
      email:
        description: 邮箱（唯一）
        type: string
      first_login:
        description: 是否首次登录
        type: boolean
      id:
        description: 主键（使用雪花算法生成，JSON序列化为字符串）
        type: integer
      level:
        description: 会员等级：normal, silver, gold, platinum
        type: string
      open_id:
        description: 微信 OpenID（仅微信登录用户有值）
        type: string
      phone:
        description: 手机号（唯一，可为空）
        type: string
      points:
        description: 积分
        type: integer
      real_name:
        description: 真实姓名
        type: string
      role:
        description: 角色：user, admin
        type: string
      status:
        description: 状态：active, blocked
        type: string
      total_spend:
        description: 总消费
        type: number
      updated_at:
        description: 更新时间
        type: string
      username:
        description: 用户名（唯一）
        type: string
    type: object
  service.AddUserRequest:
    properties:
      email:
        type: string
      phone:
        type: string
      real_name:
        type: string
      role:
        type: string
      username:
        maxLength: 20
        minLength: 2
        type: string
    required:
    - email
    - username
    type: object
...
  service.DeleteUsersRequest:
    properties:
      user_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - user_ids
    type: object
...
  service.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  service.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  service.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      real_name:
        type: string
      username:
        maxLength: 20
        minLength: 2
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  service.UpdateUserRequest:
    properties:
      avatar:
        type: string
      balance:
        type: number
      email:
        type: string
      level:
        type: string
      phone:
        type: string
      points:
        type: integer
      real_name:
        type: string
      role:
        type: string
      status:
        type: string
      total_spend:
        type: number
      username:
        maxLength: 20
        minLength: 2
        type: string
    required:
    - email
    - username
    type: object
  service.WeChatLoginRequest:
    properties:
      avatar:
        type: string
      code:
        type: string
      nickname:
        type: string
    required:
    - code
    type: object
  utils.ErrorInfo:
    properties:
      code:
        description: 错误代码
        type: string
      message:
        description: 错误消息
        type: string
    type: object
  utils.PageInfo:
    properties:
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  utils.PageResponse:
    properties:
      data: {}
      page:
        $ref: '#/definitions/utils.PageInfo'
      success:
        type: boolean
    type: object
  utils.Response:
    properties:
      data:
        description: 返回的数据
      error:
        allOf:
        - $ref: '#/definitions/utils.ErrorInfo'
        description: 错误信息
      message:
        description: 提示消息
        type: string
      success:
        description: 是否成功
        type: boolean
    type: object
host: 127.0.0.1:18080
info:
  contact:
    email: support@gohotel.com
    name: API Support
  description: 这是一个酒店预订管理系统的 RESTful API 服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 酒店管理系统 API
  version: "1.0"
paths:
...
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 管理员获取所有用户列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 邮箱
        in: query
        name: email
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 真实姓名
        in: query
        name: real_name
        type: string
      - description: 角色
        in: query
        name: role
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PageResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户列表（管理员）
      tags:
      - 管理员
  /api/admin/users/{id}:
    get:
      consumes:
      - application/json
      description: 管理员根据用户ID获取用户详细信息
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取用户详情（管理员）
      tags:
      - 管理员
    post:
      consumes:
      - application/json
      description: 管理员更新用户资料
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - 管理员
  /api/admin/users/batch:
    post:
      consumes:
      - application/json
      description: 管理员批量删除用户账户
      parameters:
      - description: 删除用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.DeleteUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 批量删除用户
      tags:
      - 管理员
  /api/admin/users/user:
    post:
      consumes:
      - application/json
      description: 管理员添加新的用户账户，默认密码为yumi123456
      parameters:
      - description: 管理员信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 添加用户
      tags:
      - 管理员
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口，返回用户信息和 JWT token
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 新用户注册接口
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: 用户注册
      tags:
      - 认证
  /api/auth/wechat-login:
    post:
      consumes:
      - application/json
      description: 微信小程序一键登录接口
      parameters:
      - description: 微信登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.WeChatLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: 微信一键登录
      tags:
      - 认证
...
  /api/users/password:
    post:
      consumes:
      - application/json
      description: 修改当前登录用户的密码
      parameters:
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 用户
  /api/users/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的个人信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取个人信息
      tags:
      - 用户
    post:
      consumes:
      - application/json
      description: 更新当前登录用户的个人信息
      parameters:
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新个人信息
      tags:
      - 用户
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}" 格式的 JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"